backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.0
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
python-multipart==0.0.9
pymongo==4.7.2
motor==3.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.7.1
